rtp update - call get_holdover from main script
